02-VisionApp/CODE/API-ExtractInfo/requirements.txt
flask
openai
pydantic
python-dotenv